requirements.txt
pandas
requests
psycopg2
SQLAlchemy
mlb-statsapi
